SOS-813 Support multi-contact actions

+++ portlets/contacts-portlet/docroot/WEB-INF/src/com/liferay/contacts/contactscenter/portlet/ContactsCenterPortlet.java
@@ -16,17 +16,21 @@ package com.liferay.contacts.contactscenter.portlet;
 
 import com.liferay.contacts.util.ContactsUtil;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
+import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.notifications.ChannelHubManagerUtil;
 import com.liferay.portal.kernel.notifications.NotificationEvent;
 import com.liferay.portal.kernel.notifications.NotificationEventFactoryUtil;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
 import com.liferay.portal.kernel.portlet.LiferayWindowState;
 import com.liferay.portal.kernel.servlet.ServletResponseUtil;
+import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.model.Group;
@@ -44,6 +48,8 @@ import com.liferay.portlet.social.service.SocialRelationLocalServiceUtil;
 import com.liferay.portlet.social.service.SocialRequestInterpreterLocalServiceUtil;
 import com.liferay.portlet.social.service.SocialRequestLocalServiceUtil;
 import com.liferay.util.bridges.mvc.MVCPortlet;
+
+import java.io.IOException;
 
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -71,27 +77,38 @@ public class ContactsCenterPortlet extends MVCPortlet {
 		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
 			WebKeys.THEME_DISPLAY);
 
-		long userId = ParamUtil.getLong(actionRequest, "userId");
+		long[] userIds = getUserIds(actionRequest);
+
 		int type = ParamUtil.getInteger(actionRequest, "type");
 
-		boolean blocked = SocialRelationLocalServiceUtil.hasRelation(
-			userId, themeDisplay.getUserId(),
-			SocialRelationConstants.TYPE_UNI_ENEMY);
+		for (long userId : userIds) {
+			try {
+				if (userId == themeDisplay.getUserId()) {
+					continue;
+				}
 
-		if (type == SocialRelationConstants.TYPE_UNI_ENEMY) {
-			SocialRelationLocalServiceUtil.deleteRelations(
-				themeDisplay.getUserId(), userId);
-		}
-		else if (blocked) {
-			return;
-		}
-
-		SocialRelationLocalServiceUtil.addRelation(
-			themeDisplay.getUserId(), userId, type);
-
-		if (blocked) {
-			SocialRelationLocalServiceUtil.addRelation(
-				userId, themeDisplay.getUserId(), type);
+				boolean blocked = SocialRelationLocalServiceUtil.hasRelation(
+					userId, themeDisplay.getUserId(),
+					SocialRelationConstants.TYPE_UNI_ENEMY);
+	
+				if (type == SocialRelationConstants.TYPE_UNI_ENEMY) {
+					SocialRelationLocalServiceUtil.deleteRelations(
+						themeDisplay.getUserId(), userId);
+				}
+				else if (blocked) {
+					continue;
+				}
+	
+				SocialRelationLocalServiceUtil.addRelation(
+					themeDisplay.getUserId(), userId, type);
+	
+				if (blocked) {
+					SocialRelationLocalServiceUtil.addRelation(
+						userId, themeDisplay.getUserId(), type);
+				}
+			}
+			catch (PortalException pe) {
+			}
 		}
 	}
 
@@ -102,11 +119,22 @@ public class ContactsCenterPortlet extends MVCPortlet {
 		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
 			WebKeys.THEME_DISPLAY);
 
-		long userId = ParamUtil.getLong(actionRequest, "userId");
+		long[] userIds = getUserIds(actionRequest);
+
 		int type = ParamUtil.getInteger(actionRequest, "type");
 
-		SocialRelationLocalServiceUtil.deleteRelation(
-			themeDisplay.getUserId(), userId, type);
+		for (long userId : userIds) {
+			try {
+				if (userId == themeDisplay.getUserId()) {
+					continue;
+				}
+
+				SocialRelationLocalServiceUtil.deleteRelation(
+					themeDisplay.getUserId(), userId, type);
+			}
+			catch (PortalException pe) {
+			}
+		}
 	}
 
 	public void exportVCard(
@@ -151,6 +179,106 @@ public class ContactsCenterPortlet extends MVCPortlet {
 		ServletResponseUtil.sendFile(
 			request, response, "vcards.vcf",
 			vCards.getBytes(StringPool.UTF8), "text/x-vcard; charset=UTF-8");
+	}
+
+	public void getContact(
+			ResourceRequest resourceRequest, ResourceResponse resourceResponse)
+		throws Exception {
+
+		long userId = ParamUtil.getLong(resourceRequest, "userId");
+
+		JSONObject jsonObject = JSONFactoryUtil.createJSONObject();
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)resourceRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		try {
+			User user = UserLocalServiceUtil.getUser(userId);
+
+			LiferayPortletResponse liferayPortletResponse =
+				(LiferayPortletResponse)resourceResponse;
+
+			boolean viewRelationActions = true;
+
+			if (SocialRelationLocalServiceUtil.hasRelation(
+					userId, themeDisplay.getUserId(),
+					SocialRelationConstants.TYPE_UNI_ENEMY)) {
+
+				viewRelationActions = false;
+			}
+			else if (SocialRelationLocalServiceUtil.hasRelation(
+						themeDisplay.getUserId(), userId,
+						SocialRelationConstants.TYPE_UNI_ENEMY)) {
+
+				viewRelationActions = false;
+			}
+
+			boolean block = SocialRelationLocalServiceUtil.hasRelation(
+				themeDisplay.getUserId(), userId,
+				SocialRelationConstants.TYPE_UNI_ENEMY);
+
+			boolean connectionRequested =
+				viewRelationActions &&
+				SocialRequestLocalServiceUtil.hasRequest(
+					themeDisplay.getUserId(), User.class.getName(),
+					themeDisplay.getUserId(),
+					SocialRelationConstants.TYPE_BI_CONNECTION, userId,
+					SocialRequestConstants.STATUS_PENDING);
+
+			boolean connected =
+				!connectionRequested &&
+				viewRelationActions &&
+				SocialRelationLocalServiceUtil.hasRelation(
+					themeDisplay.getUserId(), userId,
+					SocialRelationConstants.TYPE_BI_CONNECTION);
+
+			boolean following =
+				viewRelationActions &&
+				SocialRelationLocalServiceUtil.hasRelation(
+					themeDisplay.getUserId(), userId,
+					SocialRelationConstants.TYPE_UNI_FOLLOWER);
+
+			JSONObject userJSONObject = JSONFactoryUtil.createJSONObject();
+
+			userJSONObject.put("block", block);
+			userJSONObject.put("connectionRequested", connectionRequested);
+			userJSONObject.put("connected", connected);
+			userJSONObject.put("emailAddress", user.getEmailAddress());
+			userJSONObject.put("jobTitle", user.getJobTitle());
+			userJSONObject.put("firstName", user.getFirstName());
+			userJSONObject.put("following", following);
+			userJSONObject.put("fullName", user.getFullName());
+			userJSONObject.put("lastName", user.getLastName());
+			userJSONObject.put(
+				"portraitURL", user.getPortraitURL(themeDisplay));
+			userJSONObject.put("userId", String.valueOf(user.getUserId()));
+
+			PortletURL viewSummaryURL =
+				liferayPortletResponse.createRenderURL();
+
+			viewSummaryURL.setWindowState(LiferayWindowState.EXCLUSIVE);
+
+			viewSummaryURL.setParameter(
+				"mvcPath", "/contacts_center/view_resources.jsp");
+			viewSummaryURL.setParameter("userId",
+				String.valueOf(user.getUserId()));
+
+			userJSONObject.put("viewSummaryURL", viewSummaryURL.toString());
+
+			jsonObject.put("user", userJSONObject);
+			jsonObject.put("success", true);
+		}
+		catch (Exception e) {
+			jsonObject.put("success", false);
+
+			String message = LanguageUtil.get(
+				themeDisplay.getLocale(),
+				"an-error-occurred-while-retrieving-the-user-information");
+
+			jsonObject.put("message", message);
+		}
+
+		writeJSON(resourceRequest, resourceResponse, jsonObject);
 	}
 
 	public void getContacts(
@@ -218,9 +346,11 @@ public class ContactsCenterPortlet extends MVCPortlet {
 			userJSONObject.put("emailAddress", user.getEmailAddress());
 			userJSONObject.put("jobTitle", user.getJobTitle());
 			userJSONObject.put("firstName", user.getFirstName());
+			userJSONObject.put("fullName", user.getFullName());
 			userJSONObject.put("lastName", user.getLastName());
 			userJSONObject.put(
 				"portraitURL", user.getPortraitURL(themeDisplay));
+			userJSONObject.put("userId", String.valueOf(user.getUserId()));
 
 			PortletURL viewSummaryURL =
 				liferayPortletResponse.createRenderURL();
@@ -228,7 +358,7 @@ public class ContactsCenterPortlet extends MVCPortlet {
 			viewSummaryURL.setWindowState(LiferayWindowState.EXCLUSIVE);
 
 			viewSummaryURL.setParameter(
-				"mvcPath", "/contacts_center/view_user.jsp");
+				"mvcPath", "/contacts_center/view_resources.jsp");
 			viewSummaryURL.setParameter(
 				"userId", String.valueOf(user.getUserId()));
 
@@ -249,21 +379,34 @@ public class ContactsCenterPortlet extends MVCPortlet {
 		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
 			WebKeys.THEME_DISPLAY);
 
-		long userId = ParamUtil.getLong(actionRequest, "userId");
+		long[] userIds = getUserIds(actionRequest);
+
 		int type = ParamUtil.getInteger(actionRequest, "type");
 
-		if (SocialRelationLocalServiceUtil.hasRelation(
-				userId, themeDisplay.getUserId(),
-				SocialRelationConstants.TYPE_UNI_ENEMY)) {
+		for (long userId : userIds) {
+			try {
+				if (userId == themeDisplay.getUserId()) {
+					continue;
+				}
 
-			return;
+				if (SocialRelationLocalServiceUtil.hasRelation(
+						userId, themeDisplay.getUserId(),
+						SocialRelationConstants.TYPE_UNI_ENEMY)) {
+
+					continue;
+				}
+
+				SocialRequest socialRequest =
+					SocialRequestLocalServiceUtil.addRequest(
+						themeDisplay.getUserId(), 0, User.class.getName(),
+						themeDisplay.getUserId(), type, StringPool.BLANK,
+						userId);
+
+				sendNotificationEvent(socialRequest, themeDisplay);
+			}
+			catch (PortalException pe) {
+			}
 		}
-
-		SocialRequest socialRequest = SocialRequestLocalServiceUtil.addRequest(
-			themeDisplay.getUserId(), 0, User.class.getName(),
-			themeDisplay.getUserId(), type, StringPool.BLANK, userId);
-
-		sendNotificationEvent(socialRequest, themeDisplay);
 	}
 
 	@Override
@@ -280,6 +423,9 @@ public class ContactsCenterPortlet extends MVCPortlet {
 			else if (id.equals("exportVCards")) {
 				exportVCards(resourceRequest, resourceResponse);
 			}
+			else if (id.equals("getContact")) {
+				getContact(resourceRequest, resourceResponse);
+			}
 			else if (id.equals("getContacts")) {
 				getContacts(resourceRequest, resourceResponse);
 			}
@@ -290,6 +436,30 @@ public class ContactsCenterPortlet extends MVCPortlet {
 		catch (Exception e) {
 			throw new PortletException(e);
 		}
+	}
+
+	public void processAction(
+			ActionRequest actionRequest, ActionResponse actionResponse)
+		throws IOException, PortletException {
+
+		try {
+			String cmd = ParamUtil.getString(actionRequest, Constants.CMD);
+
+			if (cmd.equals("addSocialRelation")) {
+				addSocialRelation(actionRequest, actionResponse);
+			}
+			else if (cmd.equals("deleteSocialRelation")) {
+				deleteSocialRelation(actionRequest, actionResponse);
+			}
+			else if (cmd.equals("requestSocialRelation")) {
+				requestSocialRelation(actionRequest, actionResponse);
+			}
+		}
+		catch (Exception e) {
+			throw new PortletException(e);
+		}
+
+		sendRedirect(actionRequest, actionResponse);
 	}
 
 	public void updateSocialRequest(
@@ -321,6 +491,21 @@ public class ContactsCenterPortlet extends MVCPortlet {
 		ChannelHubManagerUtil.confirmDelivery(
 			themeDisplay.getCompanyId(), themeDisplay.getUserId(),
 			notificationEventUuid, false);
+	}
+
+	protected long[] getUserIds(ActionRequest actionRequest) {
+		long[] userIds;
+
+		long userId = ParamUtil.getLong(actionRequest, "userId", 0);
+
+		if (userId > 0) {
+			userIds = new long[] {userId};
+		}
+		else {
+			userIds = StringUtil.split(
+				ParamUtil.getString(actionRequest, "userIds"), 0L);
+		}
+		return userIds;
 	}
 
 	protected void sendNotificationEvent(
